+++ pkg/frontend/transport/handler_test.go
@@ -33,6 +33,7 @@ func TestWriteError(t *testing.T) {
 		{http.StatusGatewayTimeout, context.DeadlineExceeded},
 		{StatusClientClosedRequest, context.Canceled},
 		{http.StatusBadRequest, httpgrpc.Errorf(http.StatusBadRequest, "")},
+		{http.StatusRequestEntityTooLarge, errors.New("http: request body too large")},
 	} {
 		t.Run(test.err.Error(), func(t *testing.T) {
 			w := httptest.NewRecorder()
@@ -43,34 +44,203 @@ func TestWriteError(t *testing.T) {
 }
 
 func TestHandler_ServeHTTP(t *testing.T) {
+	roundTripper := roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+		return &http.Response{
+			StatusCode: http.StatusOK,
+			Body:       io.NopCloser(strings.NewReader("{}")),
+		}, nil
+	})
+	userID := "12345"
 	for _, tt := range []struct {
-		name            string
-		cfg             HandlerConfig
-		expectedMetrics int
+		name                       string
+		cfg                        HandlerConfig
+		expectedMetrics            int
+		roundTripperFunc           roundTripperFunc
+		additionalMetricsCheckFunc func(h *Handler)
 	}{
 		{
-			name:            "test handler with stats enabled",
-			cfg:             HandlerConfig{QueryStatsEnabled: true},
-			expectedMetrics: 3,
+			name:             "test handler with stats enabled",
+			cfg:              HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics:  3,
+			roundTripperFunc: roundTripper,
 		},
 		{
-			name:            "test handler with stats disabled",
-			cfg:             HandlerConfig{QueryStatsEnabled: false},
-			expectedMetrics: 0,
+			name:             "test handler with stats disabled",
+			cfg:              HandlerConfig{QueryStatsEnabled: false},
+			expectedMetrics:  0,
+			roundTripperFunc: roundTripper,
+		},
+		{
+			name:            "test handler with reasonResponseTooLarge",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusRequestEntityTooLarge,
+					Body:       io.NopCloser(strings.NewReader("{}")),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonRequestBodySizeExceeded, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonTooManyRequests",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusTooManyRequests,
+					Body:       io.NopCloser(strings.NewReader("{}")),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonTooManyRequests, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonTooManySamples",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitTooManySamples)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonTooManySamples, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonTooLongRange",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitTimeRangeExceeded)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonTimeRangeExceeded, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonSeriesFetched",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitSeriesFetched)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonSeriesFetched, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonChunksFetched",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitChunksFetched)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonChunksFetched, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonChunkBytesFetched",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitChunkBytesFetched)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonChunkBytesFetched, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonDataBytesFetched",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitDataBytesFetched)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonDataBytesFetched, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonSeriesLimitStoreGateway",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitSeriesStoreGateway)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonSeriesLimitStoreGateway, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonChunksLimitStoreGateway",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitChunksStoreGateway)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonChunksLimitStoreGateway, userID))
+				assert.Equal(t, float64(1), v)
+			},
+		},
+		{
+			name:            "test handler with reasonBytesLimitStoreGateway",
+			cfg:             HandlerConfig{QueryStatsEnabled: true},
+			expectedMetrics: 3,
+			roundTripperFunc: roundTripperFunc(func(req *http.Request) (*http.Response, error) {
+				return &http.Response{
+					StatusCode: http.StatusUnprocessableEntity,
+					Body:       io.NopCloser(strings.NewReader(limitBytesStoreGateway)),
+				}, nil
+			}),
+			additionalMetricsCheckFunc: func(h *Handler) {
+				v := promtest.ToFloat64(h.rejectedQueries.WithLabelValues(reasonBytesLimitStoreGateway, userID))
+				assert.Equal(t, float64(1), v)
+			},
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
-			roundTripper := roundTripperFunc(func(req *http.Request) (*http.Response, error) {
-				return &http.Response{
-					StatusCode: http.StatusOK,
-					Body:       io.NopCloser(strings.NewReader("{}")),
-				}, nil
-			})
-
 			reg := prometheus.NewPedanticRegistry()
 			handler := NewHandler(tt.cfg, roundTripper, log.NewNopLogger(), reg)
 
-			ctx := user.InjectOrgID(context.Background(), "12345")
+			ctx := user.InjectOrgID(context.Background(), userID)
 			req := httptest.NewRequest("GET", "/", nil)
 			req = req.WithContext(ctx)
 			resp := httptest.NewRecorder()